safety: common speed mismatch test (#2296)
* common test

* better place

* clean up

* these comments are useful
