+++ opendbc/safety/tests/test_rivian.py
@@ -1,11 +1,10 @@
 #!/usr/bin/env python3
 import unittest
-import numpy as np
 
 from opendbc.car.structs import CarParams
 from opendbc.safety.tests.libsafety import libsafety_py
 import opendbc.safety.tests.common as common
-from opendbc.safety.tests.common import CANPackerPanda, MAX_SPEED_DELTA
+from opendbc.safety.tests.common import CANPackerPanda
 from opendbc.car.rivian.values import RivianSafetyFlags
 from opendbc.car.rivian.riviancan import checksum as _checksum
 
@@ -58,6 +57,7 @@ class TestRivianSafetyBase(common.PandaCarSafetyTest, common.DriverTorqueSteerin
     return self.packer.make_can_msg_panda("ESP_Status", 0, values, fix_checksum=checksum)
 
   def _speed_msg_2(self, speed, quality_flag=True):
+    # Rivian has a dynamic max torque limit based on speed, so it checks two sources
     return self._user_gas_msg(0, speed, quality_flag)
 
   def _user_brake_msg(self, brake):
@@ -108,20 +108,6 @@ class TestRivianSafetyBase(common.PandaCarSafetyTest, common.DriverTorqueSteerin
         to_push[0].data[0] = 0xff
         self.assertFalse(self._rx(to_push))
         self.assertFalse(self.safety.get_controls_allowed())
-
-  def test_rx_hook_speed_mismatch(self):
-    # TODO: this can be a common test w/ Ford
-    # Rivian has a dynamic max torque limit based on speed, so it checks two sources
-    for speed in np.arange(0, 40, 0.5):
-      for speed_delta in np.arange(-5, 5, 0.1):
-        speed_2 = round(max(speed + speed_delta, 0), 1)
-        # Set controls allowed in between rx since first message can reset it
-        self._rx(self._speed_msg(speed))
-        self.safety.set_controls_allowed(True)
-        self._rx(self._speed_msg_2(speed_2))
-
-        within_delta = abs(speed - speed_2) <= MAX_SPEED_DELTA
-        self.assertEqual(self.safety.get_controls_allowed(), within_delta)
 
 
 class TestRivianStockSafety(TestRivianSafetyBase):
